🐛 change production vector index
The current production vector index was used in both dev and prod.
This leads to cluttering of the vector (with dev and prod data) and results in
empty cards showing up under similar packs (see: #1200). This was fixed by using a different one for dev instead.
The ultimate solution however is to change the production vector index to an entirely new one, thus eliminating
the chance of cluttering even when running from previous versions (older commits).